New Feature: Block Groups
Block Groups allows users to group specific blocks (per default all ores) together, so that they are being mined like a vein of the same block. This is very usefull also if you want to target specific scenarios, for example mining a tree with leaves.

Version was bumped to 2.1.0 to signify the change

Some fabric autocomplete, namely the showing of blocks in a group to remove them is not working currently, as my knowledge of Silk-Commands is too limited