implement legacy check
This checks whether Veinminer is running on server software deemed "legacy" (specifically Spigot and CraftBukkit). This is done to reduce the noise from discord users asking the same question, as because of the nature of the problem, the error messages created are very unclear and don't assist a user with non-technical background to find out what they did wrong.